Lagt til Arbeidsavtale-komponent.
