+++ src/felles-komponenter/arbeidsforhold/arbeidsforholdet.js
@@ -11,53 +11,13 @@ import Permisjoner from './permisjoner';
 import TimerTimelonnet from './timertimelonnet';
 import Utenlandsopphold from './utenlandsopphold';
 import Inntekt from './inntekt';
+import Arbeidsavtaler from './arbeidsavtaler';
 
 import { boolTilNorsk, datoDiff } from '../../utils/utils';
 
 import './arbeidsforholdet.css';
 
 const uuid = require('uuid/v4');
-
-/** Dette er komponent for iterering av arbeidsavtaler slik den leveres innenfor ett
- * arbeidsforhold
- * @param { props.avtalen } Den enkelte avtale.
- */
-function Arbeidsavtalen({ avtalen }) {
-  const {
-    arbeidstidsordning,
-    yrke,
-    beregnetAntallTimerPrUke,
-    antallTimerFraGammeltRegister,
-  } = avtalen;
-
-  return (
-    <Nav.Row>
-      <div className="arbeidsavtale">
-        <Nav.Column xs="6">
-          <dl className="arbeidsforholdet__detaljer">
-            <dt>Yrke</dt>
-            <dd>{yrke || '-'}</dd>
-            <dt>Arbeidstidsordning</dt>
-            <dd>{arbeidstidsordning}</dd>
-          </dl>
-        </Nav.Column>
-        <Nav.Column xs="6">
-          <dl className="arbeidsforholdet__detaljer">
-            <dt>Stillingsprosent</dt>
-            <dd>{beregnetAntallTimerPrUke || '-'}</dd>
-            <dt>Antall timer pr uke</dt>
-            <dd>{beregnetAntallTimerPrUke || '-'}</dd>
-            {antallTimerFraGammeltRegister && <div><dt>Antall timer fra gammelt register</dt><dd>37,5</dd></div> }
-          </dl>
-        </Nav.Column>
-      </div>
-    </Nav.Row>
-  );
-}
-
-Arbeidsavtalen.propTypes = {
-  avtalen: MPT.Arbeidsavtale.isRequired,
-};
 
 /** Dette er komponenten for ett enkelt Arbeidsforhold. Denne eksporteres ikke til omverden, men brukes
  * kun av komponenten Arbeidsforholdene.
